// src/app/model/user.model.ts
export class User{

	  _id: string;
	  password: string;
	  username: string;
	  accountNonExpired: boolean;
	  accountNonLocked: boolean;
	  credentialsNonExpired: boolean;
	  isEnabled: boolean ;
	  store_id: string ;
	  contactNumber: string ;
	  role: string ;
	  email: string ;
	  creationTime: string;

}